Prac12_MQTT/requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.10.4
altgraph==0.17.3
aniso8601==9.0.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.6.2
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
cssselect==1.2.0
cssutils==2.9.0
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.4
django-cors-headers==3.14.0
django-filter==22.1
django-ratelimit==4.0.0
django-seed==0.3.1
django-simple-tags==0.6.0
django-user-agents==0.4.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
dnspython==2.6.1
drf-yasg==1.21.4
ecdsa==0.18.0
et-xmlfile==1.1.0
executing==1.2.0
Faker==15.3.4
fastapi==0.103.2
fastjsonschema==2.16.2
flake8==6.1.0
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-MySQLdb==1.0.1
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
flatbuffers==23.5.26
fonttools==4.39.0
fqdn==1.5.1
frozenlist==1.3.3
future==0.18.2
gast==0.4.0
geoip2==4.6.0
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.8.0
httpcore==0.18.0
httpie==3.2.2
httpx==0.25.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
itypes==1.2.0
jax==0.4.12
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonify==0.5
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.0
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.12.0
kiwisolver==1.4.4
libclang==16.0.0
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
maxminddb==2.2.0
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.5
ml-dtypes==0.2.0
multidict==6.0.4
mysql-connector-python==8.0.32
mysqlclient==2.1.1
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
paho-mqtt==2.0.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
pluggy==1.0.0
premailer==3.10.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
psycopg2==2.9.5
ptw==1.0.1
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
Pygments==2.14.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
PyJWT==2.8.0
pymongo==4.6.3
PyMySQL==1.0.3
pyparsing==3.0.9
pyrsistent==0.19.3
PySide6==6.5.1
PySide6-Addons==6.5.1
PySide6-Essentials==6.5.1
PySocks==1.7.1
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.6
python-multipart==0.0.6
pytz==2022.7
pywin32==305
pywin32-ctypes==0.2.0
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
redis==4.5.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rsa==4.9
ruamel.yaml==0.17.21
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.7.2
Send2Trash==1.8.0
shiboken6==6.5.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.12
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.27.0
statsmodels==0.13.5
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
text-unidecode==1.3
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
toposort==1.7
tornado==6.2
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.8.0
tzdata==2022.7
ua-parser==0.16.1
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.13
user-agents==2.2.0
uvicorn==0.23.2
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.3
widgetsnbextension==4.0.5
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
yagmail==0.15.293
yamail==1.0.2
yarl==1.8.2
